+++ clean.sh
@@ -1,2 +1,4 @@
 #!/usr/bin/env bash
-rm -v ./glesha-ar-*.tar.gz
+rm -rfv ./dist ./build
+rm -rfv ./glesha-ar-*.tar.gz
+rm -rfv .progress.meta