+++ main.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"flag"
 	"fmt"
-	"glesha/archive"
+	"glesha/archiver"
 	"glesha/cmd"
 	"glesha/config"
 	L "glesha/logger"
@@ -25,73 +25,86 @@ func init() {
 		os.Exit(0)
 	}
 	if err != nil {
-		L.Error(err)
 		flag.Usage()
-		os.Exit(1)
+		L.Panic(err)
 	}
 	err = config.Parse(args.ConfigPath)
 
 	if err != nil {
-		L.Error(err)
-		os.Exit(1)
+		L.Panic(err)
 	}
 
 }
 
-func main() {
-	signalChannel := make(chan os.Signal, 1)
-	done := make(chan struct{})
+func handleKillSignal(signalChannel chan os.Signal, arc archiver.Archiver) {
 	signal.Notify(signalChannel, syscall.SIGINT, syscall.SIGTERM)
+	signalChannelContext, cancel := context.WithCancel(context.Background())
+	go func() {
+		<-signalChannel
+		cancel()
+	}()
+	select {
+	case <-signalChannelContext.Done():
+		L.Debug("Received kill signal, saving work")
+		err := arc.HandleKillSignal()
+		if err != nil {
+			L.Panic(err)
+		}
+		os.Exit(0)
+	}
+}
+
+func main() {
 	fmt.Println("Flags: OK")
 	args := cmd.Get()
-
+	signalChannel := make(chan os.Signal, 1)
+	defer close(signalChannel)
 	L.Debug(fmt.Sprintf("\tconfig: %s, input: %s, output: %s", args.ConfigPath, args.InputPath, args.OutputPath))
 	fmt.Println("Config File: OK")
 	configs := config.Get()
 	L.Debug(fmt.Sprintf("\tconfig::AWSAccessKey %s", configs.AWSAccessKey))
 	L.Debug(fmt.Sprintf("\tconfig::AWSBucketName %s", configs.AWSBucketName))
 	L.Debug(fmt.Sprintf("\tconfig::AWSRegion %s", configs.AWSRegion))
-	var tgzArchive archive.Archive
-	tgzArchive, err := archive.NewTarGzArchive(args.InputPath, args.OutputPath)
+	var tgzArchive archiver.Archiver
+	archiveStatusChannel := make(chan archiver.ArchiveStatus)
+	tgzArchive, err := archiver.NewTarGzArchiver(args.InputPath, args.OutputPath, archiveStatusChannel)
 	if err != nil {
-		L.Error(err)
-		os.Exit(1)
+		L.Panic(err)
 	}
-	err = tgzArchive.Plan()
-	if err != nil {
-		L.Error(err)
-		os.Exit(1)
-	}
-	fmt.Println("Plan Archive: OK")
-	filesInfo, err := tgzArchive.GetInfo()
-	if err != nil {
-		L.Error(err)
-		os.Exit(1)
-	}
-	L.Debug(fmt.Sprintf("Readable Files: %d Total Files: %d, Total Size: %d B",
-		filesInfo.ReadableFileCount,
-		filesInfo.TotalFileCount,
-		filesInfo.SizeInBytes))
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
 	go func() {
-		err = tgzArchive.Start()
+		err = tgzArchive.Plan()
 		if err != nil {
-			L.Error(err)
+			L.Panic(err)
 		}
-		done <- struct{}{}
 	}()
 
-	go func() {
-		<-signalChannel
-		cancel()
-	}()
-	select {
-	case <-done:
-		fmt.Println("Done!")
-	case <-ctx.Done():
-		L.Debug("Received kill signal, saving work")
-		tgzArchive.Pause()
-		L.Debug("Paused, exiting")
+	go handleKillSignal(signalChannel, tgzArchive)
+	L.Debug("After handleKilSignal")
+	for archiveStatus := range tgzArchive.GetStatusChannel() {
+		L.Debug(fmt.Sprintf("archiveStatus signal: %s", archiver.StatusString(archiveStatus)))
+		if archiveStatus == archiver.STATUS_PLANNED {
+			fmt.Println("Plan Archive: OK")
+			filesInfo, err := tgzArchive.GetInfo()
+			if err != nil {
+				L.Panic(err)
+			}
+			L.Debug(fmt.Sprintf("Readable Files: %d Total Files: %d, Total Size: %d B",
+				filesInfo.ReadableFileCount,
+				filesInfo.TotalFileCount,
+				filesInfo.SizeInBytes))
+			go func() {
+				err = tgzArchive.Start()
+				if err != nil {
+					L.Panic(err)
+				}
+			}()
+		}
+		if archiveStatus == archiver.STATUS_PAUSED {
+			fmt.Println("Pause Archive: OK")
+			os.Exit(0)
+		}
+		if archiveStatus == archiver.STATUS_COMPLETED {
+			fmt.Println("Create Archive: OK")
+		}
 	}
 }